comments

+++ src/bun.js/bindings/BufferEncodingType.h
@@ -5,7 +5,7 @@
 namespace WebCore {
 
 // must match src/bun.js/node/types.zig#Encoding
-enum class BufferEncodingType : uint16_t {
+enum class BufferEncodingType : uint8_t {
     utf8 = 0,
     ucs2 = 1,
     utf16le = 2,
